Update README.md
This repository includes codes and dataset used in "IRCNN：An Irregular-Time-Distanced Recurrent Convolutional Neural Network for Change Detection in Satellite Time Series ", which is submitted to IEEE Geoscience and Remote Sensing Letters. The codes and dataset are only for researches, not for commercial use. Please feel free to contact us if you have problems with the codes and dataset.